+++ cmd/app/comicvine.go
@@ -1,0 +1,9 @@
+package main
+
+/*
+type CvMedia struct {
+}
+
+func getComicVineData(mangaName string) (*CvMedia, error) {
+
+}*/
